.business/financieelplan_NL_nl/.002-draft/requirements.txt
numpy==2.1.1
pandas==2.2.3
PyYAML==6.0.2
Jinja2==3.1.4
pytest==8.3.3
matplotlib==3.9.2
